Move constraint clone tests.
